refactor: share os/arch selector normalization helpers
