+++ src/toolset/tool_request.rs
@@ -472,16 +472,9 @@ impl ToolRequest {
 
     pub(crate) fn is_os_supported(&self) -> bool {
         if let Some(os_list) = self.os() {
-            let current_os = &crate::cli::version::OS;
-            let current_arch = &crate::cli::version::ARCH;
-            let matched = os_list.iter().any(|entry| {
-                if let Some((os, arch)) = entry.split_once('/') {
-                    normalize_os(os) == current_os.as_str()
-                        && normalize_arch(arch) == current_arch.as_str()
-                } else {
-                    normalize_os(entry) == current_os.as_str()
-                }
-            });
+            let matched = os_list
+                .iter()
+                .any(|entry| crate::cli::version::os_selector_matches(entry));
             if !matched {
                 return false;
             }
@@ -664,24 +657,6 @@ fn resolve_path(p: &str, source: &ToolSource) -> PathBuf {
             .unwrap_or_else(|| PathBuf::from(".")),
     };
     base.join(p)
-}
-
-/// Normalize OS name aliases to the canonical form used by `std::env::consts::OS`.
-fn normalize_os(os: &str) -> &str {
-    match os {
-        "darwin" | "macos" => "macos",
-        "windows" | "win" => "windows",
-        other => other,
-    }
-}
-
-/// Normalize architecture name aliases to the canonical form used by `cli::version::ARCH`.
-fn normalize_arch(arch: &str) -> &str {
-    match arch {
-        "x86_64" | "amd64" | "x64" => "x64",
-        "aarch64" | "arm64" => "arm64",
-        other => other,
-    }
 }
 
 /// subtracts sub from orig and removes suffix
@@ -1050,27 +1025,5 @@ mod tests {
         assert!(version_sub("latest", "1").is_err());
         assert!(version_sub("1.2.3", "x").is_err());
         assert!(version_sub("", "1").is_err());
-    }
-
-    #[test]
-    fn test_normalize_os() {
-        use super::normalize_os;
-        assert_eq!(normalize_os("macos"), "macos");
-        assert_eq!(normalize_os("darwin"), "macos");
-        assert_eq!(normalize_os("linux"), "linux");
-        assert_eq!(normalize_os("windows"), "windows");
-        assert_eq!(normalize_os("win"), "windows");
-        assert_eq!(normalize_os("freebsd"), "freebsd");
-    }
-
-    #[test]
-    fn test_normalize_arch() {
-        use super::normalize_arch;
-        assert_eq!(normalize_arch("arm64"), "arm64");
-        assert_eq!(normalize_arch("aarch64"), "arm64");
-        assert_eq!(normalize_arch("x64"), "x64");
-        assert_eq!(normalize_arch("x86_64"), "x64");
-        assert_eq!(normalize_arch("amd64"), "x64");
-        assert_eq!(normalize_arch("riscv64"), "riscv64");
     }
 }
